+++ CHANGELOG.md
@@ -10,6 +10,8 @@ Releases prior to 7.0 has been removed from this file to declutter search result
 
 ### Fixed
 
+- cli: Fixed crash when running `schema wipe` on empty or in-memory database.
+- database: Fixed `dipdup_wipe` function not dropping all user-defined objects in some cases.
 - database: Fix exception when creating connections with aiosqlite==0.22.0.
 - evm.node: Retry JSON-RPC requests on "invalid block range params" error.
 - substrate.subscan: Pass `X-API-Key` header when `api_key` is configured.